+++ source/DSCResources/DSC_SqlTraceFlag/DSC_SqlTraceFlag.psm1
@@ -47,15 +47,42 @@ function Get-TargetResource
     if ($sqlManagement)
     {
         $databaseEngineService = $sqlManagement.Services |
-            Where-Object -FilterScript { $PSItem.Name -eq $serviceNames.SQLEngineName }
+            Where-Object -FilterScript {
+                $_.Name -eq $serviceNames.SQLEngineName
+            }
 
         if ($databaseEngineService)
         {
-            $traceFlags = $databaseEngineService.StartupParameters.Split(';') |
-                Where-Object -FilterScript { $PSItem -like '-T*' } |
-                ForEach-Object {
-                    $PSItem.TrimStart('-T')
+            Write-Debug -Message (
+                $script:localizedData.DebugParsingStartupParameters -f $MyInvocation.MyCommand, $databaseEngineService.StartupParameters
+            )
+
+            $startupParameterValues = $databaseEngineService.StartupParameters.Split(';')
+
+            $startupParameterTraceFlagValues = $startupParameterValues |
+                Where-Object -FilterScript {
+                    $_ -match '^-T\d+'
                 }
+
+            $traceFlags = @()
+
+            if ($startupParameterTraceFlagValues)
+            {
+                Write-Debug -Message (
+                    $script:localizedData.DebugFoundTraceFlags -f $MyInvocation.MyCommand, ($startupParameterTraceFlagValues -join ', ')
+                )
+
+                $traceFlags = @(
+                    $startupParameterTraceFlagValues |
+                        ForEach-Object {
+                            $_.TrimStart('-T')
+                        }
+                )
+           }
+           else
+           {
+               Write-Debug -Message ($script:localizedData.DebugNoTraceFlags -f $MyInvocation.MyCommand)
+           }
         }
         else
         {
@@ -69,12 +96,17 @@ function Get-TargetResource
         New-InvalidOperationException -Message $errorMessage
     }
 
+    Write-Debug -Message (
+        $script:localizedData.DebugReturningTraceFlags -f $MyInvocation.MyCommand, ($traceFlags -join ', ')
+    )
+
     return @{
         ServerName          = $ServerName
         InstanceName        = $InstanceName
-        TraceFlags          = $traceFlags
-        TraceFlagsToInclude = $null
-        TraceFlagsToExclude = $null
+        TraceFlags          = [System.UInt32[]] $traceFlags
+        TraceFlagsToInclude = [System.UInt32[]] @()
+        TraceFlagsToExclude = [System.UInt32[]] @()
+        ClearAllTraceFlags  = $false
         RestartService      = $null
         RestartTimeout      = $null
     }
@@ -103,6 +135,9 @@ function Get-TargetResource
     .PARAMETER TraceFlagsToExclude
         The TraceFlags the SQL server engine startup parameters should exclude.
         This parameter can not be used together with TraceFlags.
+
+    .PARAMETER ClearAllTraceFlags
+        Specifies that there should be no trace flags set on the instance.
 
     .PARAMETER RestartService
         If set, the sql server instance gets a reset after setting parameters.
@@ -141,7 +176,11 @@ function Set-TargetResource
 
         [Parameter()]
         [System.Boolean]
-        $RestartService = $false,
+        $ClearAllTraceFlags,
+
+        [Parameter()]
+        [System.Boolean]
+        $RestartService,
 
         [Parameter()]
         [System.UInt32]
@@ -161,7 +200,8 @@ function Set-TargetResource
             'TraceFlags'
         )
         MutuallyExclusiveList2 = @(
-            'TraceFlagsToInclude', 'TraceFlagsToExclude'
+            'TraceFlagsToInclude',
+            'TraceFlagsToExclude'
         )
     }
 
@@ -172,57 +212,66 @@ function Set-TargetResource
         InstanceName = $InstanceName
     }
 
-    $wishTraceFlags = [System.Collections.ArrayList]::new()
+    $desiredTraceFlags = [System.Collections.ArrayList]::new()
 
     if ($PSBoundParameters.ContainsKey('TraceFlags'))
     {
         if ($null -ne $TraceFlags)
         {
-            $wishTraceFlags.AddRange($TraceFlags)
+            $desiredTraceFlags.AddRange(@($TraceFlags))
         }
     }
-    else
+    elseif ($PSBoundParameters.ContainsKey('TraceFlagsToInclude') -or $PSBoundParameters.ContainsKey('TraceFlagsToExclude'))
     {
         $getTargetResourceResult = Get-TargetResource @getTargetResourceParameters
 
         if ($null -ne $getTargetResourceResult.TraceFlags)
         {
-            $wishTraceFlags.AddRange($getTargetResourceResult.TraceFlags)
+            $desiredTraceFlags.AddRange(@($getTargetResourceResult.TraceFlags))
         }
 
         if ($PSBoundParameters.ContainsKey('TraceFlagsToInclude'))
         {
-            foreach ($traceFlagToInclude in $TraceFlagsToInclude)
+            foreach ($currentTraceFlagToInclude in $TraceFlagsToInclude)
             {
-                if ($getTargetResourceResult.TraceFlags -notcontains $traceFlagToInclude)
+                if ($desiredTraceFlags -notcontains $currentTraceFlagToInclude)
                 {
-                    $wishTraceFlags.Add($traceFlagToInclude)
+                    $desiredTraceFlags.Add($currentTraceFlagToInclude)
                 }
             }
         }
 
         if ($PSBoundParameters.ContainsKey('TraceFlagsToExclude'))
         {
-            foreach ($traceFlagToExclude in $TraceFlagsToExclude)
+            foreach ($currentTraceFlagToExclude in $TraceFlagsToExclude)
             {
-                if ($getTargetResourceResult.TraceFlags -contains $traceFlagToExclude)
+                if ($desiredTraceFlags -contains $currentTraceFlagToExclude)
                 {
-                    $wishTraceFlags.Remove([string]$traceFlagToExclude)
+                    $desiredTraceFlags.Remove($currentTraceFlagToExclude)
                 }
             }
         }
     }
+    elseif ($PSBoundParameters.ContainsKey('ClearAllTraceFlags'))
+    {
+        Write-Verbose -Message $script:localizedData.ClearingAllTraceFlags
+
+        $desiredTraceFlags = @()
+    }
+    else
+    {
+        Write-Verbose -Message $script:localizedData.NoTraceFlagParameter
+
+        return
+    }
 
     # Add '-T' dash to flag.
-    $traceFlagList = $wishTraceFlags |
-        ForEach-Object {
-            "-T$PSItem"
-        }
-
-    if ($traceFlagList -eq '')
-    {
-        $traceFlagList = $null
-    }
+    $startupParameterTraceFlagValues = @(
+        $desiredTraceFlags |
+            ForEach-Object {
+                '-T{0}' -f $_
+            }
+    )
 
     $serviceNames = Get-SqlServiceName -InstanceName $InstanceName
 
@@ -231,33 +280,39 @@ function Set-TargetResource
     if ($sqlManagement)
     {
         $databaseEngineService = $sqlManagement.Services |
-            Where-Object -FilterScript { $PSItem.Name -eq $serviceNames.SQLEngineName }
+            Where-Object -FilterScript {
+                $_.Name -eq $serviceNames.SQLEngineName
+            }
 
         if ($databaseEngineService)
         {
             # Extract startup parameters.
-            [System.Collections.ArrayList] $parameterList = $databaseEngineService.StartupParameters.Split(';')
+            $currentStartupParameters = $databaseEngineService.StartupParameters.Split(';')
 
-            # Removing flags that are not wanted
-            foreach ($parameter in $databaseEngineService.StartupParameters.Split(';'))
+            [System.Collections.ArrayList] $parameterList = $currentStartupParameters
+
+            # Remove all current trace flags
+            foreach ($parameter in $currentStartupParameters)
             {
-                if ($parameter -like '-T*' -and $parameter -notin $traceFlagList)
+                if ($parameter -match '^-T\d+')
                 {
                     $parameterList.Remove($parameter) | Out-Null
                 }
             }
 
-            # Add missing flags.
-            foreach ($flag in $traceFlagList)
+            # Set all desired trace flags
+            foreach ($desiredTraceFlag in $startupParameterTraceFlagValues)
             {
-                if ($flag -notin $parameterList)
-                {
-                    $parameterList.Add($flag) | Out-Null
-                }
+                $parameterList.Add($desiredTraceFlag) | Out-Null
             }
 
-            # Merge flags back into startup parameters.
+            # Merge parameter list back into startup parameters.
             $databaseEngineService.StartupParameters = $parameterList -join ';'
+
+            Write-Debug -Message (
+                $script:localizedData.DebugSetStartupParameters -f $MyInvocation.MyCommand, $databaseEngineService.StartupParameters
+            )
+
             $databaseEngineService.Alter()
 
             if ($PSBoundParameters.ContainsKey('RestartService'))
@@ -295,6 +350,9 @@ function Set-TargetResource
         The TraceFlags the SQL server engine startup parameters should exclude.
         This parameter can not be used together with TraceFlags.
 
+    .PARAMETER ClearAllTraceFlags
+        Specifies that there should be no trace flags set on the instance.
+
     .PARAMETER RestartService
         If set, the sql server instance gets a reset after setting parameters.
         after restart the sql server agent is in the original state as before restart.
@@ -305,7 +363,7 @@ function Set-TargetResource
 #>
 function Test-TargetResource
 {
-    [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute('SqlServerDsc.AnalyzerRules\Measure-CommandsNeededToLoadSMO', '', Justification='The command Import-SQLPSMOdule is called when Get-TargetResource is called')]
+    [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute('SqlServerDsc.AnalyzerRules\Measure-CommandsNeededToLoadSMO', '', Justification='The command Import-SQLPSModule is called when Get-TargetResource is called')]
     [CmdletBinding()]
     [OutputType([System.Boolean])]
     param
@@ -321,20 +379,24 @@ function Test-TargetResource
         $InstanceName,
 
         [Parameter()]
-        [System.Uint32[]]
+        [System.UInt32[]]
         $TraceFlags,
 
         [Parameter()]
-        [System.Uint32[]]
+        [System.UInt32[]]
         $TraceFlagsToInclude,
 
         [Parameter()]
-        [System.Uint32[]]
+        [System.UInt32[]]
         $TraceFlagsToExclude,
 
         [Parameter()]
         [System.Boolean]
-        $RestartService = $false,
+        $ClearAllTraceFlags,
+
+        [Parameter()]
+        [System.Boolean]
+        $RestartService,
 
         [Parameter()]
         [System.UInt32]
@@ -351,7 +413,22 @@ function Test-TargetResource
             'TraceFlags'
         )
         MutuallyExclusiveList2 = @(
-            'TraceFlagsToInclude', 'TraceFlagsToExclude'
+            'TraceFlagsToInclude',
+            'TraceFlagsToExclude'
+        )
+    }
+
+    Assert-BoundParameter @assertBoundParameterParameters
+
+    $assertBoundParameterParameters = @{
+        BoundParameterList     = $PSBoundParameters
+        MutuallyExclusiveList1 = @(
+            'ClearAllTraceFlags'
+        )
+        MutuallyExclusiveList2 = @(
+            'TraceFlags'
+            'TraceFlagsToInclude',
+            'TraceFlagsToExclude'
         )
     }
 
@@ -366,77 +443,92 @@ function Test-TargetResource
 
     $isInDesiredState = $true
 
-    if ($PSBoundParameters.ContainsKey('TraceFlags'))
+    if ($PSBoundParameters.ContainsKey('ClearAllTraceFlags') -and $ClearAllTraceFlags)
     {
-        if ($TraceFlags.Length -eq 0)
+        if ($getTargetResourceResult.TraceFlags.Count -gt 0)
         {
-            if ($getTargetResourceResult.TraceFlags.Count -gt 0)
-            {
-                $isInDesiredState = $false
-            }
-        }
-        else
-        {
-            $reference = [System.Collections.ArrayList]::new()
-
-            if ($null -ne $getTargetResourceResult.TraceFlags)
-            {
-                $reference.AddRange($getTargetResourceResult.TraceFlags)
-            }
-
-            $difference = [System.Collections.ArrayList]::new()
-
-            if ($null -ne $TraceFlags)
-            {
-                $difference.AddRange($TraceFlags)
-            }
-
-            # Compare $TraceFlags to the Actual TraceFlags ($getTargetResourceResult.TraceFlags) to see if they contain the same values.
-            $nullIfTheSame = Compare-Object -ReferenceObject $reference -DifferenceObject $difference
-            if ($null -ne $nullIfTheSame)
-            {
-                Write-Verbose -Message (
-                    $script:localizedData.DesiredTraceFlagNotPresent `
-                        -f $($TraceFlags -join ','), $($getTargetResourceResult.TraceFlags -join ',')
+            Write-Verbose -Message (
+                $script:localizedData.ClearNotInDesiredState -f @(
+                    ($getTargetResourceResult.TraceFlags -join ', ')
                 )
+            )
 
-                $isInDesiredState = $false
-            }
+            $isInDesiredState = $false
+        }
+    }
+    elseif ($PSBoundParameters.ContainsKey('TraceFlags'))
+    {
+        $currentStateTraceFlags = [System.Collections.ArrayList]::new()
+
+        if (-not [System.String]::IsNullOrEmpty($getTargetResourceResult.TraceFlags))
+        {
+            $currentStateTraceFlags.AddRange(@($getTargetResourceResult.TraceFlags))
+        }
+
+        $desiredStateTraceFlags = [System.Collections.ArrayList]::new()
+
+        if (-not [System.String]::IsNullOrEmpty($TraceFlags))
+        {
+            $desiredStateTraceFlags.AddRange(@($TraceFlags))
+        }
+
+        # Returns $null if desired state and current state is the same.
+        $compareObjectResult = Compare-Object -ReferenceObject $currentStateTraceFlags -DifferenceObject $desiredStateTraceFlags
+
+        if ($null -ne $compareObjectResult)
+        {
+            $isInDesiredState = $false
+
+            Write-Verbose -Message (
+                $script:localizedData.NotInDesiredState -f @(
+                    ($desiredStateTraceFlags -join ', '),
+                    ($currentStateTraceFlags -join ', ')
+                )
+            )
         }
     }
     else
     {
         if ($PSBoundParameters.ContainsKey('TraceFlagsToInclude'))
         {
-            foreach ($traceFlagToInclude in $TraceFlagsToInclude)
+            foreach ($currentTraceFlagToInclude in $TraceFlagsToInclude)
             {
-                if ($getTargetResourceResult.TraceFlags -notcontains $traceFlagToInclude)
+                if ($getTargetResourceResult.TraceFlags -notcontains $currentTraceFlagToInclude)
                 {
                     Write-Verbose -Message (
                         $script:localizedData.TraceFlagNotPresent `
-                            -f $traceFlagToInclude
+                            -f $currentTraceFlagToInclude
                     )
 
                     $isInDesiredState = $false
+
+                    break
                 }
             }
         }
 
         if ($PSBoundParameters.ContainsKey('TraceFlagsToExclude'))
         {
-            foreach ($traceFlagToExclude in $TraceFlagsToExclude)
+            foreach ($currentTraceFlagToExclude in $TraceFlagsToExclude)
             {
-                if ($getTargetResourceResult.TraceFlags -contains $traceFlagToExclude)
+                if ($getTargetResourceResult.TraceFlags -contains $currentTraceFlagToExclude)
                 {
                     Write-Verbose -Message (
                         $script:localizedData.TraceFlagPresent `
-                            -f $traceFlagToExclude
+                            -f $currentTraceFlagToExclude
                     )
 
                     $isInDesiredState = $false
+
+                    break
                 }
             }
         }
+    }
+
+    if ($isInDesiredState)
+    {
+        Write-Verbose -Message $script:localizedData.InDesiredState
     }
 
     return $isInDesiredState
@@ -447,21 +539,21 @@ function Test-TargetResource
         This function returns the serviceNames of an sql instance.
 
     .PARAMETER InstanceName
-        The name of the SQL instance of whoose service names are beeing returned.
+        The name of the SQL instance of who's service names are being returned.
 #>
 function Get-SqlServiceName
 {
     param
     (
-        [Parameter()]
+        [Parameter(Mandatory = $true)]
         [System.String]
-        $InstanceName = 'MSSQLServer'
+        $InstanceName
     )
 
     if ($InstanceName -eq 'MSSQLSERVER')
     {
         $sqlEngineName = 'MSSQLSERVER'
-        $sqlAgentName = 'SQLSERVERAGENT'
+        $sqlAgentName = 'SQLSERVERAGENT' # cSpell: disable-line
     }
     else
     {